[refactor](be) Split parquet scalar column reader
### What problem does this PR solve?

Issue Number: close #xxx

Related PR: #xxx

Problem Summary: Move the scalar parquet column reader and its direct Arrow leaf decode path out of the complex column reader implementation so leaf decoding has its own module boundary.

### Release note

None

### Check List (For Author)

- Test: Manual test
    - Ran `./build-support/clang-format.sh be/src/format/new_parquet/column_reader.cpp be/src/format/new_parquet/scalar_column_reader.cpp be/src/format/new_parquet/scalar_column_reader.h`
    - Ran `git diff --cached --check`
- Behavior changed: No
- Does this need documentation: No
